+++ core/src/messages/api/vercel/__tests__/processStepsReal.test.ts
@@ -1,7 +1,7 @@
+import { getDefaultModels } from "@core/core-config/compat"
 import type { AgentStep } from "@core/messages/pipeline/AgentStep.types"
 import type { ModelName } from "@core/utils/spending/models.types"
-import type { LocationData } from "@core/utils/types/location.types"
-import { getDefaultModels } from "@core/core-config/compat"
+import type { LocationData } from "@lucky/shared"
 import { describe, expect, it } from "vitest"
 import responseMultiple from "../../__tests__/resources/toolResponseMultipleSteps.json"
 import responseNoToolUsed from "../../__tests__/resources/toolResponseNoToolUsed.json"
