+++ core/src/utils/json/index.ts
@@ -1,7 +1,6 @@
 /**
- * JSON utilities and database types for standalone core.
- * Replaces dependencies on @lucky/shared.
+ * JSON utilities re-exported from @lucky/shared for backward compatibility
  */
 
-export { JSONN, isJSON, show } from "./jsonParse"
+export { JSONN, isJSON, show } from "@lucky/shared"
 export type { Json, Database, Enums, Tables, TablesInsert, TablesUpdate } from "@lucky/shared"
