+++ core/tsconfig.json
@@ -36,5 +36,5 @@
     "allowSyntheticDefaultImports": true
   },
   "include": ["src/**/*.ts", "src/**/*.tsx", "src/**/*.js", "src/**/*.json", "src/**/*.d.ts", "scripts/**/*.ts"],
-  "exclude": ["node_modules", "src/examples/logging_folder", "src/utils/env.mjs"]
+  "exclude": ["node_modules", "src/examples/logging_folder", "src/utils/env.mjs", "dist"]
 }
