+++ examples/tsconfig.json
@@ -23,5 +23,5 @@
     "allowSyntheticDefaultImports": true
   },
   "include": ["**/*.ts", "**/*.tsx", "**/*.js"],
-  "exclude": ["node_modules"]
+  "exclude": ["node_modules", "dist"]
 }
