+++ package.json
@@ -23,7 +23,7 @@
     "vitest": "^3.2.4"
   },
   "overrides": {
-    "esbuild": "0.25.9",
+    "esbuild": "0.21.5",
     "tsup": "8.5.0"
   },
   "engines": {
