+++ packages/shared/package.json
@@ -38,7 +38,8 @@
   },
   "dependencies": {
     "json5": "^2.2.3",
-    "papaparse": "^5.5.3"
+    "papaparse": "^5.5.3",
+    "zod": "^3.25.76"
   },
   "devDependencies": {
     "@supabase/supabase-js": "^2.49.4",
