+++ packages/shared/src/index.ts
@@ -9,4 +9,12 @@ export * from "./utils/common/isNir"
 
 // Types
 export type { Database, Enums, Json, Tables, TablesInsert, TablesUpdate } from "./types/database.types"
-export type { StandardizedLocation } from "./types/location.types"
+export type {
+  StandardizedLocation,
+  LocationData,
+  PartialLocationData,
+  WorkflowLocationData,
+} from "./types/location.types"
+export { DataQuality, locationDataSchema, exampleLocationData } from "./types/location.types"
+export type { CodeToolName, CodeToolFailure, CodeToolSuccess, CodeToolResult } from "./types/tool.types"
+export { Tools } from "./types/tool.types"
